requirements.txt
numpy==2.0.2
pillow==10.4.0
PyAutoGUI==0.9.54